+++ lib/gemini_nano_android.dart
@@ -19,12 +19,31 @@ class GeminiNanoAndroid {
     return GeminiNanoAndroidPlatform.instance.isAvailable();
   }
 
-  /// Generates text based on a prompt using the on-device model.
+  /// Generates a list of strings based on a prompt using the on-device model.
   ///
   /// Throws a [PlatformException] if the model is not downloaded or fails.
   ///
   /// [prompt]: The input text to generate content from.
-  Future<String> generate(String prompt) {
-    return GeminiNanoAndroidPlatform.instance.generate(prompt);
+  /// [temperature]: Controls the randomness of the output. Lower values make the output more deterministic, while higher values make it more creative.
+  /// [seed]: The seed for the random number generator.
+  /// [topK]: The number of most likely tokens to consider at each step.
+  /// [candidateCount]: The number of candidate responses to generate.
+  /// [maxOutputTokens]: The maximum number of tokens to generate.
+  Future<List<String>> generate({
+    required String prompt,
+    double? temperature,
+    int? seed,
+    int? topK,
+    int? candidateCount,
+    int? maxOutputTokens,
+  }) {
+    return GeminiNanoAndroidPlatform.instance.generate(
+      prompt: prompt,
+      temperature: temperature,
+      seed: seed,
+      topK: topK,
+      candidateCount: candidateCount,
+      maxOutputTokens: maxOutputTokens,
+    );
   }
 }